Add resolver for Position entity in uni-info-watcher.
