+++ packages/uni-watcher/.eslintrc.json
@@ -28,8 +28,7 @@
         {
             "files": ["*.test.ts", "test/*.ts"],
             "rules": {
-              "no-unused-expressions": "off",
-              "no-prototype-builtins": "off"
+              "no-unused-expressions": "off"
             }
         }
     ]
