+++ packages/uni-watcher/hardhat.config.ts
@@ -7,9 +7,6 @@ const config: HardhatUserConfig = {
     compilers: [
       {
         version: '0.7.6'
-      },
-      {
-        version: '0.5.0'
       }
     ]
   },
