+++ packages/util/hardhat.config.ts
@@ -7,6 +7,9 @@ const config: HardhatUserConfig = {
     compilers: [
       {
         version: '0.7.6'
+      },
+      {
+        version: '0.5.0'
       }
     ]
   },
