+++ packages/util/test/actions.ts
@@ -1,4 +1,19 @@
-import { ethers, Contract, ContractTransaction, Signer, BigNumber } from 'ethers';
+import { ethers, Contract, ContractTransaction, Signer, BigNumber, utils } from 'ethers';
+import assert from 'assert';
+
+import {
+  abi as NFTD_ABI,
+  bytecode as NFTD_BYTECODE
+} from '@uniswap/v3-periphery/artifacts/contracts/libraries/NFTDescriptor.sol/NFTDescriptor.json';
+import {
+  abi as NFTPD_ABI,
+  bytecode as NFTPD_BYTECODE,
+  linkReferences as NFTPD_LINKREFS
+} from '@uniswap/v3-periphery/artifacts/contracts/NonfungibleTokenPositionDescriptor.sol/NonfungibleTokenPositionDescriptor.json';
+import {
+  abi as NFPM_ABI,
+  bytecode as NFPM_BYTECODE
+} from '@uniswap/v3-periphery/artifacts/contracts/NonfungiblePositionManager.sol/NonfungiblePositionManager.json';
 
 import {
   abi as TESTERC20_ABI,
@@ -8,6 +23,10 @@ import {
   abi as TESTUNISWAPV3CALLEE_ABI,
   bytecode as TESTUNISWAPV3CALLEE_BYTECODE
 } from '../artifacts/test/contracts/TestUniswapV3Callee.sol/TestUniswapV3Callee.json';
+import {
+  abi as WETH9_ABI,
+  bytecode as WETH9_BYTECODE
+} from '../artifacts/test/contracts/WETH9.sol/WETH9.json';
 
 export { abi as TESTERC20_ABI } from '../artifacts/test/contracts/TestERC20.sol/TestERC20.json';
 
@@ -56,4 +75,70 @@ export const createPool = async (
 export const initializePool = async (pool: Contract, sqrtPrice: string): Promise<void> => {
   const transaction: ContractTransaction = await pool.initialize(BigNumber.from(sqrtPrice));
   await transaction.wait();
+};
+
+export const deployWETH9Token = async (signer: Signer): Promise<string> => {
+  const WETH9 = new ethers.ContractFactory(WETH9_ABI, WETH9_BYTECODE, signer);
+  const weth9 = await WETH9.deploy();
+
+  return weth9.address;
+};
+
+// https://github.com/ethers-io/ethers.js/issues/195
+const linkLibraries = (
+  {
+    bytecode,
+    linkReferences
+  }: {
+    bytecode: string
+    linkReferences: { [fileName: string]: { [contractName: string]: { length: number; start: number }[] } }
+  },
+  libraries: { [libraryName: string]: string }): string => {
+  Object.keys(linkReferences).forEach((fileName) => {
+    Object.keys(linkReferences[fileName]).forEach((contractName) => {
+      if (!Object.prototype.hasOwnProperty.call(libraries, contractName)) {
+        throw new Error(`Missing link library name ${contractName}`);
+      }
+      const address = utils.getAddress(libraries[contractName]).toLowerCase().slice(2);
+      linkReferences[fileName][contractName].forEach(({ start: byteStart, length: byteLength }) => {
+        const start = 2 + byteStart * 2;
+        const length = byteLength * 2;
+        bytecode = bytecode
+          .slice(0, start)
+          .concat(address)
+          .concat(bytecode.slice(start + length, bytecode.length));
+      });
+    });
+  });
+  return bytecode;
+};
+
+export const deployNFPM = async (signer: Signer, factory: Contract, weth9Address: string): Promise<Contract> => {
+  // Deploy NonfungiblePositionManager.
+  // https://github.com/Uniswap/uniswap-v3-periphery/blob/main/test/shared/completeFixture.ts#L31
+  const nftDescriptorLibraryFactory = new ethers.ContractFactory(NFTD_ABI, NFTD_BYTECODE, signer);
+  const nftDescriptorLibrary = await nftDescriptorLibraryFactory.deploy();
+  assert(nftDescriptorLibrary.address, 'NFTDescriptorLibrary not deployed.');
+
+  // Linking NFTDescriptor library to NFTPD before deploying.
+  const linkedNFTPDBytecode = linkLibraries({
+    bytecode: NFTPD_BYTECODE,
+    linkReferences: NFTPD_LINKREFS
+  }, {
+    NFTDescriptor: nftDescriptorLibrary.address
+  }
+  );
+
+  const positionDescriptorFactory = new ethers.ContractFactory(
+    NFTPD_ABI,
+    linkedNFTPDBytecode,
+    signer);
+  const nftDescriptor = await positionDescriptorFactory.deploy(weth9Address);
+  assert(nftDescriptor.address, 'NFTDescriptor not deployed.');
+
+  const positionManagerFactory = new ethers.ContractFactory(
+    NFPM_ABI,
+    NFPM_BYTECODE,
+    signer);
+  return await positionManagerFactory.deploy(factory.address, weth9Address, nftDescriptor.address);
 };
